Move Json to math:: conversion to dedicated header.
